# docs/openapi-apifox.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "云舒 · 后端服务",
    "description": "AI 心理健康助手「云舒」后端服务。由 Java 版迁移而来，**接口契约与原版逐字对齐**。\n\n## 统一响应格式\n\n除 SSE 接口外，所有接口都返回同一个信封：\n\n```json\n{ \"code\": \"200\", \"msg\": \"success\", \"data\": {} }\n```\n\n- `code` 是**字符串**，`\"200\"` 表示成功，其余为业务错误码\n- **业务失败通常仍是 HTTP 200**，必须判 `code` 而不是 HTTP 状态码\n- 少数情况 HTTP 与 code 一起变：未登录 401、参数校验失败 400、越权 400\n\n## 鉴权\n\n登录 `/api/user/login` 拿到 `data.token`，之后所有需鉴权接口带：\n\n```\nAuthorization: Bearer <token>\n```\n\n**在 Apifox 里只需设置一次**：登录接口 → 后置操作 → 提取变量 `data.token`，\n然后把它绑到全局的 Bearer 认证即可（详见项目 README）。\n\n需鉴权的接口占 35/39。**无需鉴权的只有 4 个**：登录、注册、刷新令牌、登出。\n\n## 权限\n\n`userType`：`1` = 普通用户，`2` = 管理员。带 👑 的接口需要管理员，\n普通用户调用返回 **HTTP 400 + code `A0300`**（不是 403，这是原版的行为）。\n",
    "version": "0.1.0"
  },
  "paths": {
    "/api/user/login": {
      "post": {
        "tags": [
          "用户"
        ],
        "summary": "登录",
        "operationId": "login_api_user_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_LoginOut_"
                },
                "example": {
                  "code": "200",
                  "msg": "操作成功",
                  "data": {
                    "token": "<登录接口返回的 token>",
                    "refreshToken": "<登录接口返回的 refreshToken>",
                    "roleType": "1",
                    "userInfo": {
                      "id": 30,
                      "username": "demo_2026",
                      "email": "demo_2026@example.com",
                      "nickname": "演示用户",
                      "avatar": null,
                      "phone": null,
                      "gender": null,
                      "birthday": null,
                      "userType": 1,
                      "status": 1,
                      "createdAt": "2026-10-01T18:22:42",
                      "updatedAt": "2026-10-01T18:22:42",
                      "genderDisplayName": "未知",
                      "userTypeDisplayName": "普通用户",
                      "statusDisplayName": "正常",
                      "displayName": "演示用户"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [],
        "description": "\n\n🔓 **无需鉴权。**"
      }
    },
    "/api/user/add": {
      "post": {
        "tags": [
          "用户"
        ],
        "summary": "注册",
        "operationId": "register_api_user_add_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_UserDetail_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [],
        "description": "\n\n🔓 **无需鉴权。**"
      }
    },
    "/api/user/refresh": {
      "post": {
        "tags": [
          "用户"
        ],
        "summary": "刷新访问令牌",
        "operationId": "refresh_api_user_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRefreshIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_LoginOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [],
        "description": "\n\n🔓 **无需鉴权。**"
      }
    },
...
    "/api/user/logout": {
      "post": {
        "tags": [
          "用户"
        ],
        "summary": "退出登录",
        "description": "body 可选。有 refreshToken 就一并吊销。\n\n🔓 **无需鉴权。**",
        "operationId": "logout_api_user_logout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/TokenRefreshIn"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_NoneType_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
    "/api/knowledge/article/{article_id}": {
      "get": {
        "tags": [
          "知识库"
        ],
        "summary": "文章详情",
        "description": "文章不存在时返回 code=200 + data=null（**不是 404**，原版如此）。",
        "operationId": "article_detail_api_knowledge_article__article_id__get",
        "parameters": [
          {
            "name": "article_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Article Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_Union_ArticleRow__NoneType__"
                },
                "example": {
                  "code": "200",
                  "msg": "操作成功",
                  "data": {
                    "id": "db6467ec-f0e4-465f-96af-e22263176652",
                    "categoryId": 1,
                    "categoryName": "心理健康基础",
                    "title": "缓解考试焦虑的实用方法",
                    "summary": null,
                    "content": "考试焦虑是大学生常见的情绪反应。研究表明，考前进行深呼吸练习可以有效降低生理唤醒水平。具体做法：吸气四秒、屏息七秒、呼气八秒，重复四轮。此外，把复习计划拆成小块、保证睡眠、适度运动，都能显著缓解考试焦虑。如果焦虑持续两周以上并影响正常生活，建议前往学校心理健康教育中心寻求专业帮助。",
                    "coverImage": null,
                    "tags": null,
                    "authorName": null,
                    "readCount": 10,
                    "status": 1,
                    "statusText": "已发布",
                    "isFavorited": false,
                    "favoriteCount": 0,
                    "publishedAt": "2026-10-01T15:38:08",
                    "createdAt": "2026-10-01T15:38:08",
                    "updatedAt": "2026-10-02T14:47:33"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "知识库"
        ],
        "summary": "更新文章",
        "operationId": "article_update_api_knowledge_article__article_id__put",
        "parameters": [
          {
            "name": "article_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Article Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_Union_ArticleRow__NoneType__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "知识库"
        ],
        "summary": "删除文章",
        "operationId": "article_delete_api_knowledge_article__article_id__delete",
        "parameters": [
          {
            "name": "article_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Article Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_NoneType_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/psychological-chat/session/start": {
      "post": {
        "tags": [
          "心理咨询"
        ],
        "summary": "新建会话",
        "operationId": "start_session_api_psychological_chat_session_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionStartIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_SessionStartOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/psychological-chat/sessions": {
      "get": {
        "tags": [
          "心理咨询"
        ],
        "summary": "我的会话列表",
        "description": "只返回当前用户自己的会话（不带 username/nickname）。",
        "operationId": "my_sessions_api_psychological_chat_sessions_get",
        "parameters": [
          {
            "name": "currentPage",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Currentpage"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 10,
              "title": "Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_dict_"
                },
                "example": {
                  "code": "200",
                  "msg": "操作成功",
                  "data": {
                    "records": [
                      {
                        "id": 56,
                        "userId": 30,
                        "username": null,
                        "nickname": null,
                        "sessionTitle": "接口文档示例会话",
                        "startedAt": "2026-10-02T14:45:43",
                        "lastEmotionAnalysis": null,
                        "lastMessageContent": "你好呀 😊 很高兴见到你，我是你的心理健康小助手。\n\n不知道...",
                        "messageCount": 5,
                        "durationMinutes": 1
                      }
                    ],
                    "total": 1
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/psychological-chat/sessions/{session_id}": {
      "delete": {
        "tags": [
          "心理咨询"
        ],
        "summary": "删除会话",
        "description": "**幂等**：解析不出主键（尚未落库的临时会话）直接返回成功。",
        "operationId": "delete_session_api_psychological_chat_sessions__session_id__delete",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_NoneType_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/scale/my": {
      "get": {
        "tags": [
          "心理量表"
        ],
        "summary": "我的量表历史",
        "operationId": "my_history_api_scale_my_get",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_list_ScaleHistoryRow__"
                },
                "example": {
                  "code": "200",
                  "msg": "操作成功",
                  "data": []
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LoginIn": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 100,
            "title": "Username",
            "description": "用户名或邮箱"
          },
          "password": {
            "type": "string",
            "maxLength": 50,
            "minLength": 6,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginIn"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "登录接口返回的 data.token。在 Apifox 里**只填 token 本身**，会自动补 `Bearer ` 前缀。也可填 `{{token}}` 引用全局变量。"
      }
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:1236",
      "description": "本地直连后端（推荐）"
    },
    {
      "url": "http://127.0.0.1:5173",
      "description": "经前端 Vite 代理（需前端在运行）"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
}