# scripts/export_openapi.py
"""导出「Apifox 可直接导入」的 OpenAPI 文件。

FastAPI 自带的 /openapi.json 是能导入，但导入后会很难用。它缺四件事：

  1. **没有 servers（baseUrl）** —— Apifox 会要求手填服务地址
  2. **没有认证声明** —— 项目用的是自定义依赖（app/deps.py 的 get_current_user），
     不是 FastAPI 的 HTTPBearer，所以 schema 里没有 securitySchemes。
     后果是 39 个接口全部要手工加 Authorization 头 —— 这才是真正麻烦的地方。
  3. **SSE 接口被推断成 application/json** —— 实际返回 text/event-stream，
     而且有 6 类事件，schema 里一个字都没写。
  4. **没有响应示例** —— 看文档不知道 `Result` 包装长什么样。

本脚本拉取实时 schema，补齐上面四项，并跑一遍**只读接口**把真实返回体
填成 example。写操作（注册、发帖、删除等）不采集，避免污染数据。

用法：
    python scripts/export_openapi.py
    python scripts/export_openapi.py --url http://127.0.0.1:1236 --out docs/openapi-apifox.json

前置：后端需已启动（make dev 或 uvicorn）。采集示例失败不影响导出，
只是那一条没有示例。
"""

from __future__ import annotations

import argparse
import copy
import json
import re
import sys
from pathlib import Path
from typing import Any

try:
    import httpx
except ImportError:  # pragma: no cover
    sys.exit("需要 httpx：apps/api/.venv/Scripts/python.exe 里已有，用它跑本脚本")

ROOT = Path(__file__).resolve().parent.parent

# ---------------------------------------------------------------------------
# 认证：这几个路径在 app/deps.py 的 PUBLIC_PATHS 里，或在路由上就没挂鉴权依赖，
# 实测确认无令牌可访问（/api/user/logout 的 handler 也不依赖 get_current_user）。
# ---------------------------------------------------------------------------
PUBLIC_OPERATIONS = {
    ("POST", "/api/user/login"),
    ("POST", "/api/user/add"),
    ("POST", "/api/user/refresh"),
    ("POST", "/api/user/logout"),
}

# 需要管理员（userType == 2）的接口：路径里带 admin
ADMIN_HINT = "👑 需要管理员（userType=2），普通用户调用返回 HTTP 400 + code A0300。"

TAG_DESCRIPTIONS = {
    "用户": "注册、登录、令牌刷新、当前用户、改密、登出。",
    "知识库": "文章与分类。列表/详情在登录后可见，增删改与重建索引需要管理员。",
    "心理咨询": "会话管理与**流式对话（SSE）**。stream 接口返回 text/event-stream，见该接口说明。",
    "情绪日记": "用户端每日记录（同一天重复提交为**全量覆盖**语义），管理端分页查询。",
    "心理量表": "题库、提交、个人趋势，以及管理端记录查询。",
    "危机预警": "危机工单的分页、待处理数、处置与求助资源。**全部需要管理员**。",
    "数据看板": "管理端聚合统计（趋势、分布、活跃度）。",
    "文件": "文件上传，返回可访问的 URL。",
}

# SSE 真实事件序列（源码 app/routers/chat.py 的 _event_stream）
SSE_DESCRIPTION = """\
流式对话。返回 `text/event-stream`，用 POST 发起。

⚠️ FastAPI 无法描述 SSE，schema 里原本写的是 application/json，**这是错的**。

**事件序列**（严格按此顺序，括号表示可能出现 0 次）：

```
crisis?  →  agent  →  sources?  →  message*  →  (done | error)
```

| 事件 | data 结构 |
|---|---|
| `crisis` | `Result.ok({ level, title, subtitle, helplines, disclaimer })` —— 仅在检测到危机时**最先**下发，卡片内容来自固定常量，不依赖模型输出 |
| `agent` | `Result.ok({...})` —— 意图路由结果，告诉前端这一轮由哪类 Agent 应答 |
| `sources` | `Result.ok({ sources: [...] })` —— RAG 命中的知识库片段（未命中则不发） |
| `message` | `Result.ok({ content, type })` —— AI 回复的**逐段**片段，出现 0..N 次 |
| `done` | **裸 `{}`** —— 恒定成功收尾，注意**没有**被 Result 包装 |
| `error` | `Result.error("500", msg, null)` —— 失败路径，替代后续内容 |

> 契约文档只记了 4 类（crisis/message/done/error），**漏了 agent 与 sources**。
> 前端 consultation.vue 是认识这两类的，以本表为准。

**在 Apifox 里调试**：SSE 不是普通 JSON 响应，Apifox 的「返回响应」面板看不到
增量内容。建议用「运行 → 控制台」或 curl 观察：

```bash
curl -N -X POST http://127.0.0.1:1236/api/psychological-chat/stream \\
  -H "Authorization: Bearer <token>" -H "Content-Type: application/json" \\
  -d '{"sessionId":"<会话ID>","userMessage":"我最近有点焦虑"}'
```
"""

INFO_DESCRIPTION = """\
AI 心理健康助手「云舒」后端服务。由 Java 版迁移而来，**接口契约与原版逐字对齐**。

## 统一响应格式

除 SSE 接口外，所有接口都返回同一个信封：

```json
{ "code": "200", "msg": "success", "data": {} }
```

- `code` 是**字符串**，`"200"` 表示成功，其余为业务错误码
- **业务失败通常仍是 HTTP 200**，必须判 `code` 而不是 HTTP 状态码
- 少数情况 HTTP 与 code 一起变：未登录 401、参数校验失败 400、越权 400

## 鉴权

登录 `/api/user/login` 拿到 `data.token`，之后所有需鉴权接口带：

```
Authorization: Bearer <token>
```

**在 Apifox 里只需设置一次**：登录接口 → 后置操作 → 提取变量 `data.token`，
然后把它绑到全局的 Bearer 认证即可（详见项目 README）。

需鉴权的接口占 35/39。**无需鉴权的只有 4 个**：登录、注册、刷新令牌、登出。

## 权限

`userType`：`1` = 普通用户，`2` = 管理员。带 👑 的接口需要管理员，
普通用户调用返回 **HTTP 400 + code `A0300`**（不是 403，这是原版的行为）。
"""


def _sanitize(value: Any) -> Any:
    """把示例里的真实令牌换成占位符 —— 文档不该带着有效凭证流转。"""
    if isinstance(value, dict):
        out = {}
        for k, v in value.items():
            if k in ("token", "refreshToken", "accessToken") and isinstance(v, str) and v:
                out[k] = f"<登录接口返回的 {k}>"
            else:
                out[k] = _sanitize(v)
        return out
    if isinstance(value, list):
        return [_sanitize(v) for v in value]
    return value


def _set_example(op: dict, value: Any, media: str = "application/json") -> None:
    if value is None:
        return
    resp = op.get("responses", {}).get("200")
    if not resp:
        return
    resp.setdefault("content", {}).setdefault(media, {})
    resp["content"][media]["example"] = _sanitize(value)


class ExampleCollector:
    """跑一遍只读接口，把真实返回体采集下来当示例。

    只碰 GET 与登录，不采集任何写操作 —— 文档生成不该改数据。
    """

    def __init__(self, base_url: str, user: tuple[str, str], admin: tuple[str, str]) -> None:
        self.client = httpx.Client(base_url=base_url, timeout=30)
        self.creds = {"user": user, "admin": admin}
        self.tokens: dict[str, str] = {}
        self.ctx: dict[str, Any] = {}
        self.login_bodies: dict[str, Any] = {}

    def login(self) -> None:
        for role, (username, password) in self.creds.items():
            try:
                r = self.client.post(
                    "/api/user/login", json={"username": username, "password": password}
                )
                body = r.json()
                tok = (body.get("data") or {}).get("token")
                if tok:
                    self.tokens[role] = tok
                    self.login_bodies[role] = body
                    print(f"  ✓ {role} 登录成功（{username}）")
                else:
                    print(f"  ✗ {role} 登录失败：{body.get('msg')}")
            except Exception as exc:  # noqa: BLE001
                print(f"  ✗ {role} 登录异常：{exc}")

    def _get(self, path: str, role: str = "user", **kw) -> Any:
        tok = self.tokens.get(role)
        if not tok:
            return None
        try:
            r = self.client.get(path, headers={"Authorization": f"Bearer {tok}"}, **kw)
            if r.status_code != 200:
                return None
            return r.json()
        except Exception:  # noqa: BLE001
            return None

    def resolve_ids(self) -> None:
        """取真实 id，供路径参数接口使用。"""
        page = self._get("/api/knowledge/article/page")
        arts = ((page or {}).get("data") or {}).get("records") or []
        if arts:
            self.ctx["article_id"] = arts[0].get("id")

        sess = self._get("/api/psychological-chat/sessions")
        data = (sess or {}).get("data")
        rows = data.get("records") if isinstance(data, dict) else data
        if rows:
            # ⚠️ 列表接口只返回数字 `id`，没有 sessionId 字段。
            #    而 /stream 的 StreamIn.sessionId 是 **str**，传数字会被
            #    参数校验拦下（返回 HTTP 200 + code 400，很容易误当成成功）。
            #    路径参数接口倒是两种写法都收。
            sid = (rows[0] or {}).get("sessionId") or (rows[0] or {}).get("id")
            self.ctx["session_id"] = str(sid)
        else:
            # 演示账号往往没有历史会话，而 /sessions/{id}/messages 这类接口
            # 没会话就没法给示例。新建一个即可 —— 该接口只建行 + 跑规则检测，
            # 不调用 AI，开销极小。
            self.ctx["session_id"] = self._start_session()

        crisis = self._get("/api/admin/crisis/page", role="admin")
        crows = ((crisis or {}).get("data") or {}).get("records") or []
        if crows:
            self.ctx["event_id"] = crows[0].get("id")

        print(f"  解析到的上下文: {self.ctx or '（为空，相关接口将没有示例）'}")

    def _start_session(self) -> str | None:
        tok = self.tokens.get("user")
        if not tok:
            return None
        try:
            r = self.client.post(
                "/api/psychological-chat/session/start",
                headers={"Authorization": f"Bearer {tok}"},
                json={"sessionTitle": "接口文档示例会话", "initialMessage": "你好"},
            )
            return ((r.json().get("data") or {}).get("sessionId"))
        except Exception:  # noqa: BLE001
            return None

    def collect(self) -> dict[tuple[str, str], Any]:
        """返回 {(method, path): 真实响应体}。"""
        targets: list[tuple[str, str, str, dict]] = [
            ("GET", "/api/user/current", "user", {}),
            ("GET", "/api/emotion-diary/my", "user", {}),
            ("GET", "/api/scale/questions", "user", {}),
            ("GET", "/api/scale/my", "user", {}),
            ("GET", "/api/scale/my/latest", "user", {}),
            ("GET", "/api/scale/trend", "user", {}),
            ("GET", "/api/knowledge/category/tree", "user", {}),
            ("GET", "/api/knowledge/article/page", "user", {}),
            ("GET", "/api/psychological-chat/agents", "user", {}),
            ("GET", "/api/psychological-chat/sessions", "user", {}),
            ("GET", "/api/admin/crisis/page", "admin", {}),
            ("GET", "/api/admin/crisis/pending/count", "admin", {}),
            ("GET", "/api/admin/crisis/resources", "admin", {}),
            ("GET", "/api/data-analytics/overview", "admin", {}),
            ("GET", "/api/emotion-diary/admin/page", "admin", {}),
            ("GET", "/api/scale/admin/page", "admin", {}),
            ("GET", "/api/psychological-chat/admin/sessions", "admin", {}),
        ]
        # 带路径参数的
        if self.ctx.get("article_id"):
            targets.append(
                ("GET", "/api/knowledge/article/{article_id}", "user",
                 {"article_id": self.ctx["article_id"]})
            )
        if self.ctx.get("session_id"):
            targets.append(
                ("GET", "/api/psychological-chat/sessions/{session_id}/messages", "user",
                 {"session_id": self.ctx["session_id"]})
            )
            targets.append(
                ("GET", "/api/psychological-chat/session/{session_id}/emotion", "user",
                 {"session_id": self.ctx["session_id"]})
            )

        out: dict[tuple[str, str], Any] = {}
        # 登录响应体其实在 login() 里已经拿到了，别漏（它是最重要的示例，
        # 因为 Apifox 要照着它配 `data.token` 的提取规则）
        if self.login_bodies.get("user"):
            out[("POST", "/api/user/login")] = self.login_bodies["user"]
        for method, path, role, ph in targets:
            url = path
            for k, v in ph.items():
                url = url.replace("{" + k + "}", str(v))
            body = self._get(url, role=role)
            if body is not None:
                out[(method, path)] = body
        return out

    def collect_sse(self, max_lines: int = 800) -> str | None:
        """抓一段**完整**的 SSE 报文（会真实调用一次 AI）。

        要点：
        · 空行是 SSE 的事件分隔符，**必须保留**，不能过滤掉
        · 读到 done / error 才算完整，否则示例会缺收尾事件
        """
        tok = self.tokens.get("user")
        sid = self.ctx.get("session_id")
        if not tok or not sid:
            return None
        try:
            lines: list[str] = []
            ended = False
            with self.client.stream(
                "POST",
                "/api/psychological-chat/stream",
                headers={"Authorization": f"Bearer {tok}"},
                json={"sessionId": sid, "userMessage": "你好"},
            ) as r:
                if r.status_code != 200:
                    return None
                for raw in r.iter_lines():  # 不过滤空行
                    lines.append(raw)
                    if raw.startswith(("event: done", "event: error")):
                        ended = True
                    elif ended and raw.startswith("data:"):
                        break
                    if len(lines) >= max_lines:
                        break
            if not lines:
                return None
            first = next((ln for ln in lines if ln.strip()), "")
            if not first.startswith("event:"):
                # 本项目把参数校验错误也包成 HTTP 200 + code 400，
                # 所以不能只看状态码 —— 首行不是 event: 就说明压根没进流式逻辑。
                print(f"    （流式接口未返回 SSE，首行为 {first[:80]!r}，丢弃该示例）")
                return None
            text = "\n".join(lines)
            if not ended:
                text += "\n\n: —— 示例已截断（原流更长，末尾应还有 event: done）——"
            return text
        except Exception:  # noqa: BLE001
            return None


def build_enriched(spec: dict, examples: dict, sse_sample: str | None) -> dict:
    spec = copy.deepcopy(spec)

    # —— info ——
    spec["info"]["title"] = "云舒 · 后端服务"
    spec["info"]["description"] = INFO_DESCRIPTION
    spec["info"]["version"] = "0.1.0"

    # —— servers：导入 Apifox 时勾「导入 Servers 为环境」就会自动建环境 ——
    spec["servers"] = [
        {"url": "http://127.0.0.1:1236", "description": "本地直连后端（推荐）"},
        {"url": "http://127.0.0.1:5173", "description": "经前端 Vite 代理（需前端在运行）"},
    ]

    # —— 认证 ——
    comps = spec.setdefault("components", {})
    comps.setdefault("securitySchemes", {})["BearerAuth"] = {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": (
            "登录接口返回的 data.token。在 Apifox 里**只填 token 本身**，"
            "会自动补 `Bearer ` 前缀。也可填 `{{token}}` 引用全局变量。"
        ),
    }
    spec["security"] = [{"BearerAuth": []}]

    # —— tag 说明 ——
    # ⚠️ FastAPI 只在 operation 里写 tags，顶层**不生成** tags 数组。
    #    所以必须从各接口反推分组名，否则会得到一个空数组、Apifox 里没有分组说明。
    used: list[str] = []
    for ops in spec["paths"].values():
        for method, op in ops.items():
            if method not in ("get", "post", "put", "delete", "patch"):
                continue
            for name in op.get("tags") or []:
                if name not in used:
                    used.append(name)
    ordered = [t for t in TAG_DESCRIPTIONS if t in used] + [t for t in used if t not in TAG_DESCRIPTIONS]
    spec["tags"] = [{"name": name, "description": TAG_DESCRIPTIONS.get(name, "")} for name in ordered]

    # —— 逐接口增强 ——
    for path, ops in spec["paths"].items():
        for method, op in ops.items():
            if method not in ("get", "post", "put", "delete", "patch"):
                continue
            key = (method.upper(), path)

            # 公开接口：覆盖全局 security
            if key in PUBLIC_OPERATIONS:
                op["security"] = []
                op["description"] = (op.get("description") or "") + "\n\n🔓 **无需鉴权。**"
            elif "/admin" in path:
                op["description"] = (op.get("description") or "") + f"\n\n{ADMIN_HINT}"

            # SSE 接口单独描述
            if path.endswith("/stream"):
                op["description"] = SSE_DESCRIPTION
                op["responses"]["200"] = {
                    "description": "SSE 事件流（crisis? → agent → sources? → message* → done|error）",
                    "content": {
                        "text/event-stream": {
                            "schema": {"type": "string"},
                            "example": sse_sample
                            or "event: agent\ndata: {\"code\":\"200\",...}\n\n"
                            "event: message\ndata: {\"code\":\"200\",\"data\":{\"content\":\"你好\",\"type\":\"normal\"}}\n\n"
                            "event: done\ndata: {}\n\n",
                        }
                    },
                }
                continue

            # 真实响应示例
            if key in examples:
                _set_example(op, examples[key])

    return spec


def main() -> int:
    ap = argparse.ArgumentParser()
    ap.add_argument("--url", default="http://127.0.0.1:1236", help="后端地址")
    ap.add_argument("--out", default="docs/openapi-apifox.json", help="输出路径")
    ap.add_argument("--no-examples", action="store_true", help="跳过真实示例采集")
    ap.add_argument("--no-sse", action="store_true", help="跳过 SSE 采样（省一次 AI 调用）")
    args = ap.parse_args()

    print(f"① 拉取 schema: {args.url}/openapi.json")
    try:
        raw = httpx.get(f"{args.url}/openapi.json", timeout=15).json()
    except Exception as exc:  # noqa: BLE001
        return print(f"  ✗ 无法获取 schema：{exc}\n    后端启动了吗？") or 1
    n_ops = sum(
        1 for p in raw["paths"] for m in raw["paths"][p] if m in ("get", "post", "put", "delete")
    )
    print(f"  ✓ OpenAPI {raw.get('openapi')}，{n_ops} 个接口")

    examples: dict[tuple[str, str], Any] = {}
    sse_sample = None

    if args.no_examples:
        print("② 跳过示例采集")
    else:
        print("② 采集真实响应示例（只读接口）")
        col = ExampleCollector(args.url, ("demo_2026", "demo123456"),
                               ("diary_1790849826", "secret123"))
        col.login()
        col.resolve_ids()
        examples = col.collect()
        print(f"  ✓ 采集到 {len(examples)} 条示例")
        if not args.no_sse:
            print("③ 抓取 SSE 真实报文（会调用一次 AI）")
            sse_sample = col.collect_sse()
            print(f"  {'✓ 已抓到' if sse_sample else '✗ 未抓到（可能 AI 不可用）'}")

    print("④ 增强 schema")
    spec = build_enriched(raw, examples, sse_sample)

    out = ROOT / args.out
    out.parent.mkdir(parents=True, exist_ok=True)
    out.write_text(json.dumps(spec, ensure_ascii=False, indent=2), encoding="utf-8")
    size = out.stat().st_size
    print(f"  ✓ 已写入 {out}  （{size / 1024:.1f} KB）")

    # 自检
    print("\n⑤ 自检")
    n_sec = sum(
        1 for p in spec["paths"] for m, o in spec["paths"][p].items()
        if m in ("get", "post", "put", "delete") and o.get("security") == []
    )
    n_ex = sum(
        1 for p in spec["paths"] for m, o in spec["paths"][p].items()
        if (o.get("responses", {}).get("200", {}).get("content", {})
            .get("application/json", {}).get("example") is not None)
    )
    leaks = re.findall(r"eyJ[A-Za-z0-9_\-]{10,}", out.read_text(encoding="utf-8"))
    print(f"  securitySchemes : {'✓' if 'BearerAuth' in spec['components']['securitySchemes'] else '✗'}")
    print(f"  servers         : {len(spec['servers'])} 个")
    print(f"  公开接口覆盖     : {n_sec} 个（期望 4）")
    print(f"  带真实示例       : {n_ex} 个")
    print(f"  SSE 已文档化     : {'✓' if 'text/event-stream' in json.dumps(spec['paths']['/api/psychological-chat/stream']) else '✗'}")
    print(f"  残留真实 JWT     : {len(leaks)} 处 {'✓ 已清理' if not leaks else '✗ 需处理'}")
    return 0


if __name__ == "__main__":
    raise SystemExit(main())